refactor: move elements to UI directory
