+++ frontend/src/components/UI/Checkbox/Checkbox.module.scss
