+++ frontend/src/components/UI/Radio/Radio.jsx
@@ -1,33 +1,26 @@
 import PropTypes from 'prop-types';
-import styles from './CustomCheckbox.module.scss';
+import styles from './Radio.module.scss';
 
-function CustomCheckbox({
-  id,
-  name,
-  value,
-  formik,
-  label = '',
-  checked = false,
-}) {
+function Radio({ id, name, value, formik, label = '', checked = false }) {
   return (
-    <div className={styles.checkbox}>
+    <div className={styles.radio}>
       <input
         className={styles.input}
         id={id}
-        type="checkbox"
         name={name}
-        {...formik.getFieldProps({ name })}
+        type="radio"
         value={value}
         defaultChecked={checked}
+        onChange={formik.getFieldProps({ name }).onChange}
       />
       <label className={styles.label} htmlFor={id}>
-        {label}
+        {label && label}
       </label>
     </div>
   );
 }
 
-CustomCheckbox.propTypes = {
+Radio.propTypes = {
   formik: PropTypes.shape({
     getFieldProps: PropTypes.func.isRequired,
     errors: PropTypes.shape({}).isRequired,
@@ -35,15 +28,14 @@ CustomCheckbox.propTypes = {
   }).isRequired,
   id: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
-  value: PropTypes.string,
+  value: PropTypes.string.isRequired,
   label: PropTypes.string,
   checked: PropTypes.bool,
 };
 
-CustomCheckbox.defaultProps = {
-  value: undefined,
+Radio.defaultProps = {
   label: '',
   checked: false,
 };
 
-export default CustomCheckbox;
+export default Radio;
