+++ frontend/src/components/forms/CustomInput/CustomInput.module.scss
@@ -47,7 +47,7 @@
   display: block;
   min-height: $font-size-main-text-s;
   padding-left: 20px;
-  margin-top: 2px;
+  margin-top: 4px;
   @include render-font(
     $size: $font-size-main-text-s,
     $line-height: 1,
