+++ frontend/src/components/forms/Feedback/Request/Request.module.scss
@@ -56,7 +56,7 @@
 .inputs {
   display: flex;
   flex-direction: column;
-  row-gap: 16px;
+  row-gap: 13px;
   width: 100%;
   margin-top: 31px;
 }
@@ -80,7 +80,7 @@
   max-width: 358px;
   font-size: $font-size-main-text-s;
   line-height: $font-line-height-main-text-s;
-  margin-top: 14px;
+  margin-top: 10px;
   margin-right: 30px;
 }
